requirements.txt
attrs==23.1.0
autobahn==19.11.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
kiteconnect==4.2.0
numpy==1.24.3
pandas==2.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.1.1
pyotp==2.8.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.25.1
service-identity==21.1.0
six==1.16.0
smartapi-python==1.3.0
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
websocket-client==1.5.1
zope.interface==6.0
